+++ tests/testthat/test_spwb_day_c.R
@@ -94,6 +94,13 @@ test_that("spwb_day and spwb_day_c return the same result with sperry",{
   sd1_c <- medfate::spwb_day_c(x1, date, meteovec, latitude = 41.82592, elevation = 100, slope=0, aspect=0, modifyInput = FALSE)
   expect_equal(sd1$WaterBalance, sd1_c$WaterBalance) # Check for same output
   expect_equal(sd1$Plants, sd1_c$Plants) # Check for same output
+  expect_equal(sd1$Sunlit, sd1_c$Sunlit) # Check for same output
+  expect_equal(sd1$Shade, sd1_c$Shade) # Check for same output
+  expect_equal(sd1$PlantsInst, sd1_c$PlantsInst) # Check for same output
+  expect_equal(sd1$SunlitInst, sd1_c$SunlitInst) # Check for same output
+  expect_equal(sd1$ShadeInst, sd1_c$ShadeInst) # Check for same output
+  # expect_equal(sd1$Soil, sd1_c$Soil) # Check for same output
+  # expect_equal(sd1$EnergyBalance, sd1_c$EnergyBalance) # Check for same output
   
   # control_sperry$rhizosphereOverlap <- "partial"
   # x1 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sperry)
@@ -103,6 +110,29 @@ test_that("spwb_day and spwb_day_c return the same result with sperry",{
   # expect_equal(sd1, sd1_c) # Check for same output
 })
 
+
+test_that("spwb_day and spwb_day_c return the same result with sureau",{
+  control_sureau$subdailyResults <- TRUE
+  x1 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sureau)
+  sd1 <- medfate::spwb_day(x1, date, meteovec, latitude = 41.82592, elevation = 100, slope=0, aspect=0, modifyInput = FALSE)
+  sd1_c <- medfate::spwb_day_c(x1, date, meteovec, latitude = 41.82592, elevation = 100, slope=0, aspect=0, modifyInput = FALSE)
+  expect_equal(sd1$WaterBalance, sd1_c$WaterBalance) # Check for same output
+  expect_equal(sd1$Plants, sd1_c$Plants) # Check for same output
+  expect_equal(sd1$Sunlit, sd1_c$Sunlit) # Check for same output
+  expect_equal(sd1$Shade, sd1_c$Shade) # Check for same output
+  expect_equal(sd1$PlantsInst, sd1_c$PlantsInst) # Check for same output
+  expect_equal(sd1$SunlitInst, sd1_c$SunlitInst) # Check for same output
+  expect_equal(sd1$ShadeInst, sd1_c$ShadeInst) # Check for same output
+  # expect_equal(sd1$Soil, sd1_c$Soil) # Check for same output
+  # expect_equal(sd1$EnergyBalance, sd1_c$EnergyBalance) # Check for same output
+  
+  # control_sureau$rhizosphereOverlap <- "partial"
+  # x1 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sureau)
+  # sd1 <- medfate::spwb_day(x1, date, meteovec, latitude = 41.82592, elevation = 100, slope=0, aspect=0, modifyInput = TRUE)
+  # x1_c <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_sureau)
+  # sd1_c <- medfate::spwb_day_c(x1_c, date, meteovec, latitude = 41.82592, elevation = 100, slope=0, aspect=0, modifyInput = TRUE)
+  # expect_equal(sd1, sd1_c) # Check for same output
+})
 test_that("spwb and spwb_c return the same result with granier",{
   x1 <- spwbInput(exampleforest, examplesoil, SpParamsMED, control_granier)
   S1 <- medfate::spwb(x1, examplemeteo, latitude = 41.82592, elevation = 100, slope=0, aspect=0)